requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.27.0.post1

# Database + ORM
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

# Pydantic v2 + Settings
pydantic==2.7.0
pydantic-core==2.18.1
pydantic-settings==2.3.3

# Security + JWT
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.1
typing-extensions==4.10.0

# Testing Utilities
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
requests==2.31.0
